requirements.txt
numpy==2.4.2
pandas==2.3.3
joblib==1.5.3
scikit-learn==1.8.0
pyarrow==22.0.0
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.2.1
pymongo[srv]==3.12
certifi==2026.1.4
python-dotenv==1.2.1
Kaggle==1.8.3
pyyaml==6.0.3
gunicorn==25.1.0
httpx==0.28.1
pytest==9.0.2
ruff==0.15.2
black==26.1.0
boto3==1.42.59
python-json-logger==4.0.0